Warnings ruhiggestellt

+++ src/server/sv_init.c
@@ -177,7 +177,7 @@ void SV_SpawnServer (char *server, char *spawnpoint, server_state_t serverstate,
 
 	Com_DPrintf ("SpawnServer: %s\n",server);
 	if (sv.demofile)
-		FS_FCloseFile (sv.demofile);
+		FS_FCloseFile ((size_t)sv.demofile);
 
 	svs.spawncount++;		// any partially connected client will be
 							// restarted
